Move time to it's own package and make generic
