+++ time/time.go
@@ -1,4 +1,4 @@
-package util
+package time
 
 import (
 	"math"
@@ -15,23 +15,23 @@ const (
 	nanosecondsInMillisecond = int64(time.Millisecond / time.Nanosecond)
 )
 
-func TimeToMillis(t time.Time) int64 {
+func ToMillis(t time.Time) int64 {
 	return t.UnixNano() / nanosecondsInMillisecond
 }
 
-// TimeFromMillis is a helper to turn milliseconds -> time.Time
-func TimeFromMillis(ms int64) time.Time {
+// FromMillis is a helper to turn milliseconds -> time.Time
+func FromMillis(ms int64) time.Time {
 	return time.Unix(0, ms*nanosecondsInMillisecond)
 }
 
 // FormatTimeMillis returns a human readable version of the input time (in milliseconds).
 func FormatTimeMillis(ms int64) string {
-	return TimeFromMillis(ms).String()
+	return FromMillis(ms).String()
 }
 
 // FormatTimeModel returns a human readable version of the input time.
 func FormatTimeModel(t model.Time) string {
-	return TimeFromMillis(int64(t)).String()
+	return FromMillis(int64(t)).String()
 }
 
 // ParseTime parses the string into an int64, milliseconds since epoch.
@@ -40,10 +40,10 @@ func ParseTime(s string) (int64, error) {
 		s, ns := math.Modf(t)
 		ns = math.Round(ns*1000) / 1000
 		tm := time.Unix(int64(s), int64(ns*float64(time.Second)))
-		return TimeToMillis(tm), nil
+		return ToMillis(tm), nil
 	}
 	if t, err := time.Parse(time.RFC3339Nano, s); err == nil {
-		return TimeToMillis(t), nil
+		return ToMillis(t), nil
 	}
 	return 0, httpgrpc.Errorf(http.StatusBadRequest, "cannot parse %q to a valid timestamp", s)
 }
